Update for visibility scoping API changes in core
Note that I got rid of some of the existing policy logic (eg.
showWithFlags) because I don't know why it's there in the first place?
(Why should a post become visible to *everyone* if it has been flagged?
And what would that even do, because the discussion would still be
invisible...)

So I may have broken some edge case behaviour, but I've tested the main
functionality and it seems to be working.